requirements.txt
allure-pytest==2.14.2
Faker==37.3.0
mysql-connector-python==9.3.0
pydantic==2.11.4
pytest==8.3.5
pytest-xdist==3.6.1
python-dotenv==1.1.0
requests==2.32.3
